// src/app/models/product-interface.ts
export interface ProductInterface {
    id_producto?: string;
    nombre_producto?: string;
    tipo_producto?: string;
    descripcion?: string;
    id_sucursal?: string;
    nombre_sucursal?: string;
    ciudad_sucursal?: string;
    cantidad?: number;
    precio?: number;
    precio_vp?: number;
}